Move visualizer code to its own module
